smartsuccess-phase2/gpu-server/requirements.txt
# SmartSuccess GPU Server Requirements
# For self-hosted GPU server running Whisper + XTTS + RAG

# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0

# GPU/ML
torch>=2.1.0
torchaudio>=2.1.0

# STT - Whisper
openai-whisper>=20231117

# TTS - XTTS/Coqui
TTS>=0.22.0

# RAG & Embeddings
sentence-transformers>=2.2.2
chromadb>=0.4.22

# Document Processing
PyMuPDF>=1.23.0  # For PDF
python-docx>=1.1.0  # For DOCX

# Utilities
numpy>=1.24.0
httpx>=0.26.0
